// src/app/models/CPanelCustomerSaleReport.model.ts
export class CPanelCustomerSaleReport{

  play_master_id?: number;
  user_name?: string;
  total?: number;
  commission?: number;
  prize_value?: number;
  claimed_prize_value?: number;
  unclaimed_prize_value?: number;
  super_stockist_commission?: number;
  stockist_commission?: number;
  ntp?:number
}